+++ NEWS.md
@@ -1,5 +1,9 @@
 
-# groupdata2 1.3.0.9000
+# groupdata2 1.4.0
+
+* Breaking: In `fold()`, the `k` argument can now be a multi-element vector with one `k` (number of folds) per fold column. This functionality required a minor rewrite, why you might see interchanged fold column names in comparison to the previous versions.
+
+* Bug fix: In `fold()` and `partition()`, when specifying multiple `cat_col` columns *and* `num_col` in the same call, it would fail. This now works.
 
 # groupdata2 1.3.0